Version 5.1 - Dec. 03, 2022
system("cls") is a Windows exclusive thing. Run this code on Windows (Codeblocks or similar) to avoid getting errors!
system("cls") is a Windows exclusive thing. Run this code on Windows (Codeblocks or similar) to avoid getting errors!
system("cls") is a Windows exclusive thing. Run this code on Windows (Codeblocks or similar) to avoid getting errors!
system("cls") is a Windows exclusive thing. Run this code on Windows (Codeblocks or similar) to avoid getting errors!